Implemented a Scriptable object installer
The dependencies of the testInstaller have been subjugated to a scriptable object installer that has been linked to the scene. This object also has access to the Container and binds it's own TestInstaller.Settings to the container so that the TestInstaller can get it through injecting a TestInstaller.Settings varaible.