Improved upon entity types and included colors
Entities now use an enum that represents each type of entity that can exist. Each entity type now has it's own color, which is set in the object installer. There is a color class that is passed into the entityManager from it being binded to the container from TestInstallerSettings (the object installer).