Included game bounderies and a background
A new WindowManager will create a background and implement a gameSize class to keep the entities within a set of bounderies.

Also improved how the players are controlled. The functions need to be refactored but they are working as intended: The player cannot move past the game bounderies but their attached entities can.